+++ packages/most-buffer/src/index.js
@@ -1,1 +1,4 @@
-module.exports = require('./most-buffer.js');
+const buffer = require('./most-buffer.js');
+
+module.exports = buffer;
+module.exports.default = buffer;