+++ packages/most-buffer/src/most-buffer.d.ts
@@ -1,3 +1,4 @@
 import { Stream } from 'most';
 
-export default function buffer<T>(count: number): (stream: Stream<T>) => Stream<T[]>;
+export function buffer<T>(count: number): (stream: Stream<T>) => Stream<T[]>;
+export default buffer;