+++ packages/most-limiter/src/most-limiter.d.ts
@@ -1,3 +1,4 @@
 import { Stream } from 'most';
 
-export default function limiter<T>(interval: number, capacity?: number): (stream: Stream<T>) => Stream<T>;
+export function limiter<T>(interval: number, capacity?: number): (stream: Stream<T>) => Stream<T>;
+export default limiter;